llm_samples/requirements.txt
websockets
litellm
weaviate-cli
langchain